Migrate monetary amounts from floating-point to integer minor units
Convert all monetary values (balance, amount, fee, limits) from double/float
to integer minor units (kobo, pesewas, cents) across the full stack to
eliminate floating-point rounding errors in financial calculations.

Server (functions/index.js):
- Add toMinorUnits(), toMajorUnits(), formatMinorUnits() helpers
- Update safeAdd/safeSubtract with Math.round()
- Update verifyPayment, initializeTransaction, initiateWithdrawal to use
  minor units natively (no more * 100 / 100 conversions)
- Update MoMo API calls to convert back to major units for external API
- Update sendMoney fee/limit calculations for minor units
- Update fraud detection thresholds for minor units
- Replace all .toFixed(2) notification messages with formatMinorUnits()

Flutter models:
- WalletModel: balance, dailyLimit, monthlyLimit, dailySpent, monthlySpent
  changed from double to int with minor-unit defaults
- TransactionModel: amount, fee, convertedAmount changed from double to int;
  added displayAmount, displayFee, displayTotalAmount, displayConvertedAmount
  helpers; exchangeRate stays as double (ratio, not monetary)

Flutter services & providers:
- CurrencyService.formatAmount/formatAmountWithSeparators accept int
- PaymentService method signatures accept int amounts
- WalletService.sendMoney/addMoney accept int amounts
- MomoService methods accept int amounts
- WalletProvider balance getter returns int
- SendMoneyState amount/convertedAmount changed to int

Flutter screens:
- All screens convert user input (major units) to minor units before
  sending to backend: (double.parse(text) * 100).round()
- All display formatting converts minor to major: (amount / 100).toStringAsFixed(2)
- BalanceCard, TransactionTile accept int amounts
- Updated add_money, withdraw, send_money, confirm_send, scan_qr,
  payment_result, transaction_details screens

+++ lib/core/services/momo_service.dart
@@ -23,7 +23,7 @@ class MomoService {
   /// Request payment from user's MoMo wallet
   /// Returns referenceId to check status
   Future<MomoPaymentResult> requestToPay({
-    required double amount,
+    required int amount,
     required String phoneNumber,
     String? currency,
     String? payerMessage,
@@ -69,7 +69,7 @@ class MomoService {
 
   /// Transfer money to user's MoMo wallet
   Future<MomoPaymentResult> transfer({
-    required double amount,
+    required int amount,
     required String phoneNumber,
     String? currency,
     String? payeeNote,
